remove unneccessary casts related to millis()

+++ msrc/escHW3.cpp
@@ -32,7 +32,7 @@ void EscHW3::update()
             }
         }
     }
-    if ((uint32_t)millis() - tsEsc_ > 150)
+    if (millis() - tsEsc_ > 150)
     {
         pwm_ = 0;
         thr_ = 0;
